Change fee_total column type to DECIMAL and update button link to new book orders page

+++ DDL.sql
@@ -39,7 +39,7 @@ CREATE TABLE Members(
   address VARCHAR(145) NOT NULL,
   email VARCHAR(145), 
   phone_number char(10),
-  fee_total FLOAT NOT NULL DEFAULT 0,
+  fee_total DECIMAL(10,2) NOT NULL DEFAULT 0,
   PRIMARY KEY (member_id)
 );
